abstracted generator and selector block preview adapter functions
